requirements.txt
annotated-types==0.6.0
asgiref==3.7.2
autopep8==2.0.4
beautifulsoup4==4.12.2
dj-config-url==0.1.1
Django==4.2.8
django-cleanup==8.0.0
django-cors-headers==4.3.1
django-easy-audit==1.3.5
django-extensions==3.2.3
django-jazzmin==2.6.0
django-ninja==1.0.1
django-silk==5.0.4
ecdsa==0.18.0
future==0.18.3
gprof2dot==2022.7.29
pyasn1==0.5.1
pycodestyle==2.11.1
pydantic==2.5.2
pydantic_core==2.14.5
python-decouple==3.8
python-jose==3.3.0
rsa==4.9
six==1.16.0
soupsieve==2.5
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.8.0
tzdata==2023.3
